+++ global/type.h
@@ -34,8 +34,8 @@ typedef unsigned char u8;
 typedef short i16;
 typedef unsigned short u16;
 
-typedef long i32;
-typedef unsigned long u32;
+typedef int i32;
+typedef unsigned int u32;
 
 typedef long long i64;
 typedef unsigned long long u64;
